+++ CHANGELOG.md
@@ -9,11 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.7.0][] - 2024-04-29
 
+### Added
+
+-   `@lumx/icons`: added icons TikTok, TencentQQ, TencentVideo and Baidu
+
 ### Changed
 
 -   `@lumx/icons`: optimized all JS SVG path icons (23% reduction in size)
 -   `@lumx/icons`: removed `@mdi/js` dependency, icons are copied in the library (necessary for the optimization)
--   `@lumx/icons`: added icons TikTok, TencentQQ, TencentVideo and Baidu
 
 ## [3.6.8][] - 2024-04-18