+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     Defaults to `cover` to scale & crop the image (like before).
     Can be changed to `contain` to avoid cropping the image (aka letterboxing).
 
+### Changed
+
+-   Flag: allow any react node in label.
+
 ## [3.7.0][] - 2024-04-29
 
 ### Added